Fixed the number types for all shapes and trait

+++ exercises/scala 3/src/traitsandinheritance/Circle.scala
@@ -3,10 +3,10 @@ package traitsandinheritance
 /**
   * Created by stefanm on 22/02/2016.
   */
-class Circle(val radius: Int) extends Shape {
-  def sides(): Int = ???
+class Circle(val radius: Double) extends Shape {
+  def sides(): Double = ???
 
-  def perimeter(): Int = ???
+  def perimeter(): Double = ???
 
-  def area(): Int = ???
+  def area(): Double = ???
 }
