+++ exercises/scala 3/src/traitsandinheritance/Rectangle.scala
@@ -3,10 +3,10 @@ package traitsandinheritance
 /**
   * Created by stefanm on 22/02/2016.
   */
-class Rectangle(val x: Int, val y: Int) extends Shape {
-  def sides(): Int = ???
+class Rectangle(val x: Double, val y: Double) extends Shape {
+  def sides(): Double = ???
 
-  def perimeter(): Int = ???
+  def perimeter(): Double = ???
 
-  def area(): Int = ???
+  def area(): Double = ???
 }
