+++ exercises/scala 3/src/traitsandinheritance/Shape.scala
@@ -4,9 +4,9 @@ package traitsandinheritance
   * Created by stefanm on 22/02/2016.
   */
 trait Shape {
-  def sides(): Int
+  def sides(): Double
 
-  def perimeter(): Int
+  def perimeter(): Double
 
-  def area(): Int
+  def area(): Double
 }
